+++ plugin/src/net/groboclown/idea/p4ic/v2/server/cache/sync/ClientCacheManager.java
@@ -358,6 +358,11 @@ public class ClientCacheManager {
                 }
             }
         }
+
+        @Override
+        public boolean hasPendingUpdates() {
+            return ! state.getPendingUpdates().isEmpty();
+        }
     }
 
 }